requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==45.0.3
dill==0.4.0
distro==1.9.0
docling==2.35.0
docling-core==2.33.0
docling-ibm-models==3.4.3
docling-parse==4.0.1
easyocr==1.7.2
ebooklib==0.19
et-xmlfile==2.0.0
exceptiongroup==1.3.0
filelock==3.18.0


filetype==1.2.0
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.3
idna==3.10
imageio==2.37.0
jinja2==3.1.6
jiter==0.10.0
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
latex2mathml==3.78.0
lazy-loader==0.4
lxml==5.4.0
markdown-it-py==3.0.0
marko==2.1.3
markupsafe==3.0.2
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.18
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.6
openai==1.82.1
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
reportlab==4.2.0
packaging==25.0
pandas==2.2.3
pdfminer-six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pluggy==1.6.0
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.5
pydantic-core==2.33.2
pydantic-settings==2.9.1
pygments==2.19.1
pylatexenc==2.10
pymupdf==1.26.0
pypdfium2==4.30.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-pptx==1.0.2
pytz==2025.2
pywin32==310
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
rtree==1.4.0
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.3
semchunk==2.2.2
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sympy==1.14.0
tabulate==0.9.0
tifffile==2025.5.10
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.52.4
typer==0.15.4
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
xlsxwriter==3.2.3
